ENH: speed calculation for triplegs and positionfixes - refactor and add simple tripleg speed
