+++ trackintel/analysis/__init__.py
@@ -10,8 +10,6 @@ from .location_identification import location_identifier
 from .location_identification import pre_filter_locations
 from .location_identification import freq_method, osna_method
 
-from .speed import speed_positionfixes, mean_speed_triplegs
-
 __all__ = [
     "temporal_tracking_quality",
     "split_overlaps",
@@ -22,6 +20,4 @@ __all__ = [
     "pre_filter_locations",
     "freq_method",
     "osna_method",
-    "speed_positionfixes",
-    "mean_speed_triplegs",
 ]
